+++ Src/sort/heap_sort.c
@@ -37,7 +37,7 @@
 #include "utils.h"
 
 static void
-Heapify(byte_t* buffer, int32_t n, int32_t i, uint32_t element_size,
+Heapify(byte_t* buffer, int32_t n, int32_t i, int32_t element_size,
         bool (*compareFun)(void* first, void* second)) {
     bool continue_iterating = true;
     int32_t index = i;
@@ -48,21 +48,21 @@ Heapify(byte_t* buffer, int32_t n, int32_t i, uint32_t element_size,
         int32_t left = (2 * index) + 1;
         int32_t right = (2 * index) + 2;
 
-        bool compare_ret_value = compareFun(&elements[left * (int32_t)element_size],
-                                            &elements[largest * (int32_t)element_size]);
+        bool compare_ret_value = compareFun(&elements[left * element_size],
+                                            &elements[largest * element_size]);
 
         if ((left < n) && (compare_ret_value)) {
             largest = left;
         }
 
-        compare_ret_value = compareFun(&elements[right * (int32_t)element_size], &elements[largest * (int32_t)element_size]);
+        compare_ret_value = compareFun(&elements[right * element_size], &elements[largest * element_size]);
 
         if ((right < n) && (compare_ret_value)) {
             largest = right;
         }
 
         if (largest != index) {
-            Utils_swapElements(&elements[index * (int32_t)element_size], &elements[largest * (int32_t)element_size], element_size);
+            Utils_swapElements(&elements[index * element_size], &elements[largest * element_size], (uint32_t)element_size);
             index = largest;
         } else {
             continue_iterating = false;
@@ -71,7 +71,7 @@ Heapify(byte_t* buffer, int32_t n, int32_t i, uint32_t element_size,
 }
 
 void
-HeapSort_sort(byte_t* buffer, int32_t number_of_elements, uint32_t element_size,
+HeapSort_sort(byte_t* buffer, int32_t number_of_elements, int32_t element_size,
               bool (*compareFun)(void* first, void* second)) {
     int32_t i;
     byte_t* elements = buffer;
@@ -80,7 +80,7 @@ HeapSort_sort(byte_t* buffer, int32_t number_of_elements, uint32_t element_size,
     }
 
     for (i = number_of_elements - 1; i >= 0; --i) {
-        Utils_swapElements(&elements[0], &elements[i * (int32_t)element_size], element_size);
+        Utils_swapElements(&elements[0], &elements[i * element_size], (uint32_t)element_size);
         Heapify(elements, i, 0, element_size, compareFun);
     }
 }
